requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-allauth==0.52.0
django-bootstrap-icons==0.8.2
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-environ==0.9.0
django-import-export==3.0.2
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django_debug_toolbar==3.8.1
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
phonenumbers==8.13.4
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
